pyproject.toml
[tool.poetry]
name = "whelp-python-task"
version = "0.1.0"
description = ""
authors = ["Zakir Abdullayev <zakirabdullayev8@gmail.com>"]
readme = "README.md"
packages = [{include = "whelp_python_task"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic-settings = "^2.2.1"
peewee = "^3.17.1"
pymysql = "^1.1.0"
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
celery = "^5.3.6"
ipdata = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
